+++ src/app/components/features/recipe/RecipeList/index.ts
@@ -1,0 +1,1 @@
+export { RecipeListVertical, RecipeListHorizontal } from './RecipeList';
